demo/requirements.txt
gradio
torch
torchvision
timm
numpy
scipy
Pillow
